pyproject.toml
[tool.poetry]
name = "wave-function-collapse"
version = "0.1.0"
description = ""
authors = ["dianchia <dianrui520@hotmail.com>"]
readme = "README.md"
packages = [{include = "wave_function_collapse"}]

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.5.2"
hydra-core = "^1.3.2"
rich = "^13.6.0"

[tool.black]
line-length = 99

[tool.ruff]
line-length = 99


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
